+++ src/components/Button/ButtonInner.tsx
@@ -69,7 +69,7 @@ const ButtonInner: FC<ButtonInnerProps> = (props) => {
       )}
       <ButtonLabel>{label}</ButtonLabel>
       {variant === "brush" && <ButtonBorders background={background} />}
-      <ButtonFocusUnderline color={underlineColor} name="Underline" />
+      <ButtonFocusUnderline color={underlineColor} name="Underline1" />
     </>
   );
 };