+++ src/context/Menu/MenuProvider.tsx
@@ -1,8 +1,9 @@
-import { createContext, FC, useState } from "react";
+import { createContext, FC, useCallback, useState } from "react";
 
 type MenuContext = {
   open: boolean;
   toggleMenu: () => void;
+  closeMenu: () => void;
 };
 
 export const menuContext = createContext<MenuContext | null>(null);
@@ -10,11 +11,18 @@ export const menuContext = createContext<MenuContext | null>(null);
 export const MenuProvider: FC = ({ children }) => {
   const [open, setOpen] = useState(false);
 
+  const toggleMenu = useCallback(() => {
+    setOpen((_open) => !_open);
+  }, [setOpen]);
+
+  const closeMenu = useCallback(() => {
+    setOpen(false);
+  }, [setOpen]);
+
   const menuValue: MenuContext = {
     open,
-    toggleMenu: () => {
-      setOpen(!open);
-    },
+    toggleMenu,
+    closeMenu,
   };
 
   return (