+++ src/node-lib/cms/sanity-client/schemas/index.ts
@@ -2,3 +2,4 @@ export * from "./base";
 export * from "./webinar";
 export * from "./blog";
 export * from "./teamMember";
+export * from "./policyPage";