+++ src/node-lib/cms/types/documents.ts
@@ -42,3 +42,12 @@ export type Attachment = {
     };
   };
 };
+
+export type PolicyPage = Document & {
+  title: string;
+  slug: string;
+  lastUpdatedAt: Date;
+  bodyPortableText: PortableTextJSON;
+};
+
+export type PolicyPagePreview = Pick<PolicyPage, "title" | "slug">;