+++ src/styles/theme/default.theme.ts
@@ -45,6 +45,8 @@ const theme: OakTheme = {
     teachersLilac: "#c6d1ef",
     twilight: "#E5D1E0",
     videoBlue: "#CEE7E5",
+    // hyperlinks
+    hyperlink: "#374CF1",
   },
   contrastColors: {
     white: "black",
@@ -89,6 +91,7 @@ const theme: OakTheme = {
     teachersLilac: "black",
     twilight: "black",
     videoBlue: "black",
+    hyperlink: "white",
   },
   buttonIconBackgroundColors: {
     teachersHighlight: "teachersYellow",
@@ -131,7 +134,7 @@ const theme: OakTheme = {
         placeholder: "grey7",
         icon: "grey7",
         border: "grey8",
-        background: "grey3",
+        background: "white",
       },
       valid: {
         text: "black",